REST Backend Project-Smart ToDo API/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pymongo==4.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
